shopping_agent/requirements.txt
# Personal Shopping Agent - Requirements
# Core dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client (for testing)
httpx>=0.24.0

# Additional utilities
python-multipart>=0.0.6

# MongoDB
pymongo>=4.5.0
motor>=3.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Email validation
email-validator>=2.0.0
